search: populate Stats.Repos before applying display limit
Our stats should be calculated before applying the display limit. IE
stats should be based on the match limit (count:). This moves the
calculation of Repos to before we apply the display limit. Additionally
it now lives inside of the progress agreggator rather than in the http
handler, which is more appropriate.