Only return 401 on invalid access token
The auth middleware returned 401 errors for all other kinds of errors as well. This was very confusing to me and seems incorrect as well. I saw logs about an access token being invalid so I dug for quite a bit until I found error messages in logs that revealed the real problem: on k8s the frontend couldn't connect to the database during a rollout.