Remove graphqlbackend dependency from codemonitors stores (#28062)
This commit establishes a clean separation between the graphqlbackend layer
and the database layer for code monitors. Previously, the code monitor store
imported and reused types from graphqlbackend. This creates types specific
to the store, and converts to those types before calling the store.

This is necessary in order to move the store wrapper into a shared database
package. We don't want to make graphqlbackend a transient dependency
on every package that imports the database package.