propagate db through handlers
This propagates a database handle through our http handlers. This has
been the biggest blocker to me getting rid of global db handles
entirely, since the remaining places we use dbconn.Global are almost all
direct descendents of these handlers. I've been putting this change off
because it's very messy, and requires passing a db directly through a
bunch of functions, but I think this is the only way to move forward
here. There are a lot of circular references in our handlers which makes
it impossible to isolate this change cleanly without just doing it all
at once.

Anyways, this gets rid of one more instance of dbconn.Global, and will
enable getting rid of many more.

1 down, 37 to go.