Reorder executor docs a bit, add first stub of bare-metal docs
This started out with me fixing a bunch of typos and ended up with me
adding bullet points, reordering things to make it easier for new
readers, and adding a first version of the bare-metal docs so we know
where they would be.

I still need to dogfood the bare-metal deployment here, including the
dependency installation, so I'll probably change details.