Convert globalstatedb package to a standard store (#28118)
This converts the globalstatedb into a more standard store. Each of the
functions in the globalstatedb package are converted to methods on the
new GlobalStateStore type. This allows us to generate mocks for the store,
getting rid of the global package-level mocks, and also allows us to get rid of
a few more instances of dbconn.Global by passing in a database from a
parent context.