remove dbconn.Global from dbtesting
The dbtesting package previously would set `dbconn.Global`, then return
it from `GetDB`. Instead, this creates a global variable just for the
`dbtesting` package so it will continue to reuse the same connection
across all calls to `GetDB` (no change in behavior), except it will no
longer set `dbconn.Global`. This package is deprecated, and will soon be
removed, so I didn't feel the need to refactor this further.