fix esbuild manifest format (#28221)
* fix esbuild manifest format

In #26594 the property names of the object that is serialized to webpack.manifest.json were changed from (eg) `app.js` to `appBundle` (and so on for CSS, runtime, and react). This broke esbuild by causing it to write an incorrect webpack.manifest.json file, which is how the Go frontend prints out the right `<script>` tags for HTML pages to load the scripts. This changes the property names written by esbuild back to the same ones that the Go frontend code expects (and the same ones that Webpack writes to the file).

* Update types

Co-authored-by: Tom Ross <tom@umpox.com>