internal/extsvc/github: Reuse V3Client.GetVersion in V4Client
In this commit, we stop making a unique API request to /meta to
identify the GitHub Enterprise Server's version. It turns out that we
already have existing code to do this in the V3Client. As a result,
it's nicer to reuse the code with the trade-off of embedding the
V3Client within the V4Client.

This will not have any affect in the existing API quota usages of
the GraphQL client since the /meta endpoint is part of the REST API
itself and we continue to use the REST API with this change.