git-combine: synthesize a monorepo (#28228)
This introduces a new command "git-combine" which will combine all
upstream HEAD branches into a single repository. For each commit
upstream it will create a new commit with the file contents as a
sub directory (the remote name).

Additionally this comes with a naive "daemon" mode which just runs git
fetch --all, git-combine, git push on a timer. This is running in
production right now and the result can be seen at
https://github.com/sgtest/megarepo