search-blitz: use a simple config file for queries
The YAML file was frustrating to edit. Instead we can use a very simple
text based format. This can be extended in the future, but I really like
how simple this is. Additionally, I believe this makes it easier to copy
paste queries into sourcegraph.

This commit keeps all the internal configuration structures the same to
minimize the commit. The text file was created with:

  yq -r '.groups[0].queries[] | "# " + .name + "\n" + .query'

We lose some of the comments that existed in the old file, but they
weren't very valuable and outdated.