internal/extsvc/github: Reuse V3Client.GetVersion in V4Client (#28279)
Fixes #28276.

In this commit, we stop making a unique API request to /meta to
identify the GitHub Enterprise Server's version. It turns out that we
already have existing code to do this in the V3Client. As a result,
it's nicer to reuse the code with the trade-off of initialising the
V3Client within the V4Client when required.

This will not have any affect in the existing API quota usages of
the GraphQL client since the /meta endpoint is part of the REST API
itself and we continue to use the REST API with this change.

Link to GitHub API docs: 
https://docs.github.com/en/enterprise-server@3.3/rest/reference/meta