httpcli, actor: propagate actor on all internal requests (#28117)
This change propagates the actor in the context (from the `actor` package) on _all_ outgoing internal requests and adds middleware that services can use to pick up actors from incoming requests and add it to the request context. The middleware is currently only applied to `frontend`'s internal API handler, which already has something similar. Also adds metrics to see the number of requests that are user-authenticated vs internal-authenticated.

This more generalized actor handling replaces all existing usage of `X-Sourcegraph-User-ID`. The header is also renamed to `X-Sourcegraph-Actor-UID` to more accurately reflect `actor.UID`, though backcompat is retained mostly with the existing behaviour (though drops support for setting the ID to 0 to imply an internal user)