Remove dbtesting package (#28426)
This removes the deprecated dbtesting package. It was only used in ~40 tests, so I just finished it off by converting them all to use dbtest.NewDB(). For the most part, this was just a mechanical rename, but I've marked any additional work needed to do this or fix tests explicitly with comments. There were a few spots that depended on the global behavior of dbtesting.