gitserver: use standardized actor propagation (#28450)
Removes the custom actor propagation that gitserver and gitserver.Client uses, in favour of leveraging the standardized actor propagation introduced in #27918. There should be no behavioural changes, since this actor data is only used for observability on the gitserver side, and serverside backcompat with the old header is retained for the time being.

One particular usage of gitserver.Client, in NewGitoliteSource, required adding the actor transport manually because the external source constructors are provided a httpcli.ExternalClientFactory, which does not have the actor propagation transport set up.