internal/extsvc/github: Simplify cachedGetRepository definition + usage
The double negative in the statement `if !nocache` makes it harder to
reason about what's happening. Instead `if cache` is simpler and
requires less CPU cycles of the brain but means the same.

Additionally, in this commit, we also remove the inline closure from
the invocation of `cachedGetRepository` and make it explicit. This was
making it hard to identify at a glance that a fourth argument with
value `false` is being passed to this method call.