searcher: remove hardcoded timeout in fetch
We have a passed in context which is managed by the diskcache.Store.
Lets rely on that instead of using a hardcoded timeout. This will likely
help remove an error that occurs on large monorepos. Currently
diskcache.Store has a timeout of 10min, but here we had it set to 2min.
So we effectively never waited more than 2min. Lets be consistent and
also wait 10min.