dbtest: skip updateDSNFromEnv if PGDATASOURCE specified (#28792)
This doesn't really make sense to do, since the user has already
specified the exact string to use. We still keep the update code since
that allows us to mutate the default test DSN.

This code could be better and I will follow-up with cleanups
here. However, this fix makes it possible for me to run db tests locally
again. Some recent refactors exposed issues in this code. In particular
I use a unix socket for my postgres, and the interactions with that
caused problems.