tracking-issue: add support for OPTIONAL LABEL (#28665)
This adds a new marker, `<!-- OPTIONAL LABEL: my-label -->` that allows you to add labels on a tracking issue that _do not_ need to be present on child issues for them to be considered part of this tracking issue. This is useful for making tracking issues easier to find without adding labels to every single issue within the tracking issue.