codeintel: Use git.ListDirectoryChildren
This change is needed so that we no longer send raw git commands to
gitserver form outside of the vcs/git package.

We ended up just copying the implmentation from
lib/codeintel/pathexistnce as the vcs/git package is not visible to that
package and we don't want out internal vcs/git package to rely on
packages in codeintel/lib.