git: Apply sub-repo permissions to git.LsFiles
We apply filtering after fetching the list of files.

This is the first PR where we actually use the sub-repo client and in
order to be pragmatic we've chosen to initialise a global client in main
so that we can share cache across all calls.

This PR is intended as an example of how we can use the client to filter
git backed data.