postgresdsn: add test for envvars we generate on NixOS (#28830)
This is a test case to ensure we don't break how we connect to databases
for people using our nix hooks. On nix we setup a postgres which is
exposed via a unix socket. The test case we added includes an example
set of envvars we create and ensures we get out the correct DSN.

This recently broke for our test environments but not in this code. This
is just making sure we don't accidently break it as we confidently
refactor :)