store: move TempZipFromFiles helper to only call site
This helper was reusing the store to create a zip file on disk. This is
a bit of overkill, so this commit simplifies the function.  Additionally
there is only one call site for this helper, so we can reduce the amount
of exported API.

I also noticed one latent issue where we ignored the errors returned by
the underlying functions we are testing. We now check those errors.