search: introduce ctx to QueryToZoektQuery
We want to be able to check feature flags. Alternatively we may want to
instead communicate this in the high level structs we pass
around. However, that was a more complicated change to understand but I
am happy to follow-up and do that. For now I want feature flags to allow
gating the new lang support which is causing some issues.