sg: add a sg check command to run existing checks
Adds a new command to sg, `sg check [topic]` which will concurrently run
every scripts currently run on the CI by the checks step. This is the
first step to remove bash scripts as it opens the way to write those in
Go which is much more reliable (but not exclusively). This is also the
first step to provide better annotations on failed builds.

It merges seamlessly with the existing check scripts through a helper
that wraps them for the time being.

The subcommands are organized by usage, focusing on what kind of
activity the user is currently doing. This opens up the path to
selectively run those checks on the CI, depending on the changes in the
PR being built.

Presently, it does not run frontend linters as this set of changes does
not modify in any way the CI steps, which are to be shipped in an
ulterior PR.