frontend: Gracefully handle broken pipe error when highlighting
We currently handle a few know error types and we've received a support
reqeust from a customer that they're consistently seeing a broken pipe
error for a particular file.

This is most likely caused by a crash on syntect-server so instead of
showing an error to the user we fall back to rendering a non-highlighted
version of the file.