ssbc: Move frontend closer to EM3 designs (#28938)
This PR brings in a new header for SSBC: the wizard layout. This PR also tweaks on many little things towards feature EM3 designs. As it couldn't get much worse, I think this is a net improvement, even if there are some things that are not fully fleshed out just yet.