search: fix negative Content size
The GraphQL field `contentByteSize` is defined as `Int` which is not enough
to represent content size of very big repositories. For some customers
and for our monorepo we see negative content size on the index status
page in site admin view.

I had to make changes in frontend as well, because frontend uses a
function `prettyBytes` to format the output but unfortunately it only
takes `number` as input. Hence I wrote a little custom function.

Once this PR is in a good shape we should also convert other fields to
BigInt.