Convert PermsStore to an interface (#29211)
This converts PermsStore to an interface, adds it as a method to the new EnterpriseDB interface, and generates a mock for it. The global PermsStore mocks currently are blocking getting rid of the non-enterprise global mocks, so this pattern will allow us to mock all our stores with the same pattern.