Convert a bunch of tests using global perms mocks (#29216)
Now that we have a mockable PermsStore interface, this converts a bunch of tests that use the global perms mocks to instead use injected mocks. This is particularly exciting because the inability to mock out the perms store (because we didn't have a way to mock enterprise stores before #29204) was blocking the conversion of a bunch of these tests.