Remove unused global mock methods (#29385)
This removes a bunch of now-unused global mocks. It should be possible now to write all new tests with the new injected DB mocks, so there is no good reason to keep these around. Also, it makes it more clear how much work is left to remove the global mocks entirely.