do not exit early on partial commit search failure (#29480)
Before this commit, we were using an errgroup, which cancels its context
when any of its tasks fails. For commit search, we'd rather continue our
search, but return the error. Currently, errors are not surfaced to the
user if results are returned, so if there are more results than from the
failing repo, the user won't see that results are incomplete. That is
out of scope of this fix though.