Fix pupeteer test
Debugging the integration test revealed that for some reason the input
isn't focused and so the text never makes it to the input.

Using click instead of focus is also what `driver.replaceText` does.

But I don't know why any of my changes in this PR would affect this
test.