Remove MockRepos (#29506)
This removes the last references to the global MockRepos struct. There are a couple of mocks that lived on MockRepos that were actually mocking methods in internal/repos, so I moved those mocks to that package. Ideally, I think I'd like to merge the internal/repos package into the database repo store, since it's very confusing to have two different repo stores.