Fix routing issue by introducing a wrapper component
The issues with the previous implementation was that the route render
function wasn't run again after the search was submitted and thus the
search results component was never rendered.

This commit fixes this issue by wrapping the conditional logic in an
actual component which can properly subscribe to the query state store
and gets rerendered when a search query is present in the query (or
not).

This wasn't a problem in the past because all data was passed down via
props, guaranteeing that the parent component would re-render if
anything changes. But this is different now that we move things to
stores and directly reading data from the stores is prone to such
errors.