search: don't resolve symlinks
With this change we treat symlinks consistently for search and for
display.

Symlinks are treated as files with the relative path of the link's
target as content. This behavior matches the experience on GitHub.

Before:

We tried to resolve symlinks to their target's content to show the
content instead of the relative path in the UI. However the
code was buggy and probably never worked correctly.

Additionaly, showing the content instead of the relative path is
confusing for the user because during search we don't resolve to the
content. Hence a search might match a symlink (matching elements of the
path) but when displayed, the file match wouldn't contain a match (or
vice versa).