search: enable repository ranking by default
We currently only have reordering result by repository rank enabled on
Sourcegraph.com. This switches it on with the same queue depth we use
for sourcegraph.com. The queue depth is a heuristic we have tuned which
balances latency with the number of backends you have.

Sourcegraph.com is our biggest known cluster (by far), so this default
value is likely larger than needs be. However, since this code was
introduced we have fixed issues around not receiving any updates from a
backend. We now regularly send updates to the queue from backends, even
if no results are found. So I believe we don't need to be as
conservative with this value as in the past. So I believe this value is
safe for customers. As a follow-up we should try tuning this value
higher for sourcegraph.com and get a bit more stability in ranking.