remove deadline from searcher (#29845)
This removes the deadline parameter from search requests sent to
searcher because it is no longer necessary now that searcher streams its
results.

When making a request to searcher, we examine the context passed to the
requesting code and pull off a deadline from that context if it is set.
Then, we explicitly marshal and send that deadline as part of the search
request. This is a historical artifact of searcher being batch-only.
This made it possible for searcher to stop its search early if it knew
its context was going to be canceled soon, and send the results back to
the caller before the caller gave up on a response.

However, now that searcher streams results as it finds them, it's fine
to just stop sending results when the context is canceled.