buildchecker: return first N builds in series of failures (#29930)
In today's branch lock, because >3 failures occurred before buildchecker got a chance to run, the root cause of the incident was not tagged because we take the most recent N builds in the series of consecutive failures. This change corrects buildchecker to take the oldest N builds in a series to ensure the root cause of the incident is included.

We also refactor `checkConsecutiveFailures` into a separate file as `findConsecutiveFailures` to reflect how it is shared between 'check' and 'history'.