pass db handle to jobs at runtime
This modifies the `Run` signature on jobs to take a `db` handle instead
of a repo pager. The repo pager can be constructed from a db handle and
a set of repo options, and many of the jobs already store the repo
options on their struct.

By passing in a db handle to `Run()`, we remove all (I think) "runtime"
fields from the job structs. As a cool side effect, this means we could
theoretically fully serialize and deserialize a job tree. I don't have
any specific use cases for this in mind, but I can see some potentially
very cool things we could do there.