Address feedback
Instead of delegating to a real link, this is now using a fake link.
This means we have to handle ctrl/meta clicks and enter ourselves and
delegate to react-router accordingly.

A adjusted the CSS so that the existing :focus-visible rule (which is
using box-shadow) is visible (it wasn't).

The a text cursor is now shown when directly hovering over text.

Some "oddities" that someone might run into:

- Because ctrl+click selects the whole line in Firefox, I assume every
  ctrl+click is actually supposed to open the result in a new tab. This
  prevents using ctrl+click to select multiple ranges (shift+click still
  works though as long as some text is already selected).
- You cannot deselect text by clicking on some code outside the
  selection... it will be interpreted as clicking the link.