user invites: make it easier to test the /welcome page (#29996)
It's quite annoying today for designers, reviewers, etc. to navigate to the
`/welcome` page because you need to:

1. Add the `enablePostSignupFlow` to your site config file in the `dev-private` repo.
2. Restart your dev server, ensuring you are running `sg start dotcom` mode.
3. Clear your browser local storage (because you may have completed the welcome flow
   already and that apollo state is cached in local storage)
4. Run a GraphQL query to edit your "temporary settings" which are stored on the server
   to enable this user flow:

```
mutation {
  editTemporarySettings(settingsToEdit:"{\"signup.finishedWelcomeFlow\": false}") {
    alwaysNil
  }
}
```

Finally, you can navigate to `/welcome` to test the flow / check designs. That's insane.

So I've added a debug parameter, so all you need now is `sg start dotcom` and navigate to
`/welcome?debug=1` to get there. None of that stuff above.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>